+++ yash-semantics/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `<yash_syntax::syntax::CompoundCommand as command::Command>::execute` now
+  honors the `CaseContinuation` specified for the executed case item.
 - External dependency versions:
     - Rust 1.79.0 → 1.82.0